MISC: LIB: UTEST: Continued work on HoneydConfig - rework profile management
Restored intended workflow to the profile functions, unit tests all pass, just needs
testing before implementation.

+++ NovaTest/src/tester_HoneydConfiguration.h
@@ -105,46 +105,58 @@ TEST_F(HoneydConfigurationTest, test_Port)
 	}
 }
 
-TEST_F(HoneydConfigurationTest, test_RenameProfile)
+TEST_F(HoneydConfigurationTest, test_profile)
 {
 	//Create dummy profile
-	NodeProfile * p = new NodeProfile();
-	p->m_name = "TestProfile";
-	p->SetEthernet("Dell");
-	p->m_icmpAction = "block";
-	p->m_parentProfile = "default";
-	p->m_udpAction = "block";
-	p->m_tcpAction = "reset";
-	p->m_uptimeMax = "100";
-	p->m_uptimeMin = "10";
+	NodeProfile p;
+	p.m_name = "TestProfile";
+	p.m_distribution = 100;
+	p.m_generated = false;
+	p.m_ethernetVendors.clear();
+	pair<string, double> ethVendor;
+	ethVendor.first = "Dell";
+	ethVendor.second = 100;
+	p.m_ethernetVendors.push_back(ethVendor);
+	p.m_icmpAction = "block";
+	p.m_parentProfile = "default";
+	p.m_udpAction = "block";
+	p.m_tcpAction = "reset";
+	p.m_uptimeMax = "100";
+	p.m_uptimeMin = "10";
+	for(uint i = 0; i < 8; i++)
+	{
+		p.m_inherited[i] = false;
+	}
 
 	// Delete the profile if it already exists
-	m_config->DeleteProfile("TestProfile-renamed");
-	m_config->DeleteProfile("TestProfile");
+	if(m_config->m_profiles.keyExists("TestProfile-renamed"))
+	{
+		EXPECT_TRUE(m_config->DeleteProfile("TestProfile-renamed"));
+	}
+	if(m_config->m_profiles.keyExists("TestProfile-renamed"))
+	{
+		EXPECT_TRUE(m_config->DeleteProfile("TestProfile"));
+	}
 
 	//Test adding a profile
 	EXPECT_TRUE(m_config->AddProfile(p));
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile") != m_config->m_profiles.end());
+	EXPECT_TRUE(m_config->m_profiles.keyExists("TestProfile"));
 
 	//Test renaming a profile
 	EXPECT_TRUE(m_config->RenameProfile("TestProfile", "TestProfile-renamed"));
+	EXPECT_TRUE(m_config->m_profiles.keyExists("TestProfile-renamed"));
+	EXPECT_FALSE(m_config->m_profiles.keyExists("TestProfile"));
 
-	m_config->UpdateAllProfiles();
+	EXPECT_TRUE(m_config->AddProfile(p));
+	EXPECT_TRUE(m_config->m_profiles.keyExists("TestProfile"));
 
-	// Save and delete object
+	EXPECT_TRUE(m_config->InheritProfile("TestProfile-renamed", "TestProfile"));
 	EXPECT_TRUE(m_config->SaveAllTemplates());
-	delete m_config;
-
-	// Reload from the config file
-	m_config = new HoneydConfiguration();
-	EXPECT_TRUE(m_config != NULL);
 	EXPECT_TRUE(m_config->LoadAllTemplates());
 
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile-renamed") != m_config->m_profiles.end());
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile") == m_config->m_profiles.end());
-
-
-	EXPECT_TRUE(m_config->DeleteProfile("TestProfile-renamed"));
+	EXPECT_TRUE(m_config->DeleteProfile("TestProfile"));
+	EXPECT_FALSE(m_config->m_profiles.keyExists("TestProfile"));
+	EXPECT_FALSE(m_config->m_profiles.keyExists("TestProfile-renamed"));
 	EXPECT_TRUE(m_config->SaveAllTemplates());
 }
 
@@ -158,107 +170,4 @@ TEST_F(HoneydConfigurationTest, test_errorCases)
 	EXPECT_EQ(NULL, m_config->GetProfile("aouhaosnuheaonstuh"));
 	EXPECT_EQ(NULL, m_config->GetNode(""));
 	EXPECT_EQ(NULL, m_config->GetNode("aouhaosnuheaonstuh"));
-
-}
-
-TEST_F(HoneydConfigurationTest, test_Profile)
-{
-	//Create dummy profile
-	NodeProfile * p = new NodeProfile();
-	p->m_name = "TestProfile";
-	p->SetEthernet("Dell");
-	p->m_icmpAction = "block";
-	p->m_parentProfile = "default";
-	p->m_udpAction = "block";
-	p->m_tcpAction = "reset";
-	p->m_uptimeMax = "100";
-	p->m_uptimeMin = "10";
-
-	bool dmEn = Config::Inst()->GetIsDmEnabled();
-	Config::Inst()->SetIsDmEnabled(false);
-
-	// Delete the profile if it already exists
-	m_config->DeleteProfile("TestProfile-renamed");
-	m_config->DeleteProfile("TestProfile");
-
-	//Test adding a profile
-	EXPECT_TRUE(m_config->AddProfile(p));
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile") != m_config->m_profiles.end());
-
-	//Modify the test profile to add a second one.
-	p->m_parentProfile = "TestProfile";
-	p->m_name = "TestProfile2";
-	EXPECT_TRUE(m_config->AddProfile(p));
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile2") != m_config->m_profiles.end());
-
-	//Test renaming a profile
-	EXPECT_TRUE(m_config->RenameProfile("TestProfile2", "TestProfile-2"));
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile2") == m_config->m_profiles.end());
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile-2") != m_config->m_profiles.end());
-
-	//Test Inheriting of a profile
-	EXPECT_TRUE((m_config->m_profiles.find("TestProfile-2")->second.m_parentProfile.compare("default")));
-	EXPECT_TRUE(!(m_config->m_profiles.find("TestProfile-2")->second.m_parentProfile.compare("TestProfile")));
-	EXPECT_TRUE(m_config->InheritProfile("TestProfile-2", "default"));
-	EXPECT_TRUE(!(m_config->m_profiles.find("TestProfile-2")->second.m_parentProfile.compare("default")));
-	EXPECT_TRUE((m_config->m_profiles.find("TestProfile-2")->second.m_parentProfile.compare("TestProfile")));
-
-	//Test deleting a profile
-	EXPECT_TRUE(m_config->DeleteProfile("TestProfile"));
-	EXPECT_TRUE(m_config->m_profiles.find("TestProfile") == m_config->m_profiles.end());
-
-	Config::Inst()->SetIsDmEnabled(dmEn);
-}
-
-
-TEST_F(HoneydConfigurationTest, test_NewProfileSaving)
-{
-	EXPECT_TRUE(m_config->AddPort(1, (portProtocol)1, (portBehavior)0, "NA") == "1_TCP_block");
-	EXPECT_TRUE(m_config->AddPort(2, (portProtocol)1, (portBehavior)1, "NA") == "2_TCP_reset");
-	EXPECT_TRUE(m_config->AddPort(3, (portProtocol)1, (portBehavior)2, "NA") == "3_TCP_open");
-
-	NodeProfile * p = new NodeProfile();
-	for (int i = 0; i < INHERITED_MAX; i++)
-	{
-		p->m_inherited[i] = true;
-	}
-	p->m_name = "test";
-	p->m_parentProfile = "default";
-	p->AddPort("1_TCP_block", false, 100);
-	p->AddPort("2_TCP_reset", false, 100);
-	p->AddPort("3_TCP_open", false, 100);
-
-	EXPECT_TRUE(m_config->AddProfile(p));
-	//EXPECT_TRUE(m_config->SaveAllTemplates());
-}
-
-TEST_F(HoneydConfigurationTest, test_profileDeletion)
-{
-	NodeProfile *parent = new NodeProfile();
-	parent->SetName("parent");
-	parent->SetParentProfile("default");
-	EXPECT_TRUE(m_config->AddProfile(parent));
-
-	m_config->UpdateAllProfiles();
-
-	NodeProfile *child = new NodeProfile();
-	child->SetName("child");
-	child->SetParentProfile("parent");
-	EXPECT_TRUE(m_config->AddProfile(child));
-
-	// Save, reload
-	EXPECT_TRUE(m_config->SaveAllTemplates());
-	EXPECT_TRUE(m_config->LoadAllTemplates());
-
-	// Delete the child and make sure it worked after a reload
-	EXPECT_TRUE(m_config->DeleteProfile("child"));
-	EXPECT_TRUE(m_config->SaveAllTemplates());
-	EXPECT_TRUE(m_config->LoadAllTemplates());
-	EXPECT_TRUE(m_config->m_profiles.find("child") == m_config->m_profiles.end());
-
-	// Delete the parent and make sure it worked after a reload
-	EXPECT_TRUE(m_config->DeleteProfile("parent"));
-	EXPECT_TRUE(m_config->SaveAllTemplates());
-	EXPECT_TRUE(m_config->LoadAllTemplates());
-	EXPECT_TRUE(m_config->m_profiles.find("parent") == m_config->m_profiles.end());
 }
